src/carpool-app/backend/requirements.txt
Django
djangorestframework
django-cors-headers
phonenumbers
requests
python-dotenv